Player Pickup System Updated
Added a series of improvements to the player pickup system, including:
- Refactored Interactable.cs structure, derive new interactble objects from the abstract class Interactable.cs
