+++ Cadenza-Unity/Assets/Scripts/Interactions/Interactable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f865a996ebd1dc4429e8f6b99a0e9bf5
+guid: 4920ff14165b60843809f60cff75008f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
