+++ Cadenza-Unity/Assets/Scripts/Interactions/Pickup.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 namespace Interactions
 {
-    public class Pickup : Targetable, IInteractable
+    public class Pickup : Interactable
     {
         private Transform playerHand;
-        private Rigidbody rigidbody; 
+        private Rigidbody rigidbody;
 
         protected override void Start()
         {
@@ -16,22 +16,22 @@ namespace Interactions
             rigidbody = gameObject.AddComponent<Rigidbody>();
             rigidbody.isKinematic = true;
             rigidbody.useGravity = false;
-            
+
             playerHand = FindObjectOfType<PlayerController>().transform.Find("Camera").transform
                 .Find("PickupContainer");
 
             GameEvents.Current.ONInteract += OnInteract;
-            GameEvents.Current.OnDrop += OnDrop; 
+            GameEvents.Current.OnDrop += OnDrop;
         }
 
-        public void OnInteract(int id)
+        public override void OnInteract(int id)
         {
-            if (this.ID != id) return;
+            if (ID != id) return;
 
             rigidbody.isKinematic = true;
             rigidbody.useGravity = false;
 
-            Transform thisTransform = transform;
+            var thisTransform = transform;
             thisTransform.SetParent(playerHand);
             thisTransform.localPosition = Vector3.zero;
             thisTransform.localRotation = Quaternion.Euler(Vector3.zero);
@@ -39,11 +39,11 @@ namespace Interactions
 
         public void OnDrop(int id)
         {
-            if (this.ID != id) return;
-            
+            if (ID != id) return;
+
             transform.parent = null;
             rigidbody.isKinematic = false;
-            rigidbody.useGravity = true; 
+            rigidbody.useGravity = true;
         }
     }
 }
